A number of fixes for the upcoming release 2.0.40: - Fixed: image-related errors in PostScript output - Fixed: some paths to files generated using HTML2PS_DIR contained duplicate directory separator - Fixed: CSS cache returns references to CSS objects instead of object values (PHP 4) - Fixed: page breaks were incorrectly positioned inside tables with col/rowspans - Fixed: CSS maring / padding properties could be parsedincorrectly when values were separated with more than one space - Fixed: Capital omega symbol was not displayed in Postscript output - Fixed: notice message caused by not-perfectly-valid external file URLs - Fixed: notices/errors caused by missing images - Fixed: width issue for floating / absolute positioned items - Fixed: list number not is formatted using the same styles as list body - Fixed: background image positioning was incorrect for non-square background images - Fixed: external image-related PNG output errors - Change: documentation updates - Change: minor function naming style changes - Change: unicode-aware line breaking algorithm.

+++ branches/2.0-stable/image.class.php
@@ -37,10 +37,18 @@ class Image {
   }
 
   function sx() {
+    if (!$this->_handle) {
+      return 0;
+    };
+
     return imagesx($this->_handle);
   }
 
   function sy() {
+    if (!$this->_handle) {
+      return 0;
+    };
+
     return imagesy($this->_handle);
   }
 }
@@ -91,9 +99,13 @@ class ImageFactory {
     // register it in the cached objects array
     //
     $handle = do_image_open($filename, $type);
-    $g_image_cache[$url] =& new Image($handle,
-                                      $filename,
-                                      $type);
+    if ($handle) {
+      $g_image_cache[$url] =& new Image($handle,
+                                        $filename,
+                                        $type);
+    } else {
+      $g_image_cache[$url] = null;
+    };
     // return image
     //
     // return do_image_open($filename);
@@ -154,7 +166,9 @@ class ImageFactory {
   //
   function clear_cache() {
     foreach ($GLOBALS['g_image_cache'] as $key => $value) {
-      unlink($value->get_filename());
+      if (!is_null($value)) {
+        unlink($value->get_filename());
+      };
     };
     $g_image_cache = array();
   }
